server/pyproject.toml
[project]
name = "symbology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  # Minimal dependencies for production API deployment
  "fastapi>=0.128.0",
  "uvicorn==0.34.1",
  "pydantic==2.11.3",
  "pydantic-settings==2.9.0",
  "sqlalchemy==2.0.40",
  "psycopg2==2.9.10",
  "structlog==24.1.0",
  "python-dotenv==1.0.0",
  "uuid7==0.1.0",
  "safe-result",
  "alembic>=1.18.3",
  "ruff>=0.15.0",
  "anthropic>=0.52.0",
]

[project.optional-dependencies]
# Development dependencies for data ingestion, analysis, and LLM features
dev = [
  "edgartools>=5.14.0",
  "httpx==0.28.1",
  "requests>=2.32.3",
  "matplotlib==3.10.3",
  "seaborn==0.13.2",
  "jupyter",
  "ipywidgets",
  "click>=8.1.0",
  "rich>=13.0.0",
]

# Testing dependencies
test = ["pytest>=8.3.5", "pytest-cov>=6.1.1", "pytest-json-report>=1.5.0"]

# All development dependencies (convenience group)
all = ["symbology[dev,test]"]

# [tool.uv.workspace]
# members = ["src/python"]
[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:pydantic.*:",
  "ignore::UserWarning:pydantic.*:",
]
testpaths = ["tests"]
norecursedirs = []
markers = [
  "integration: requires a running PostgreSQL database (deselect with '-m not integration')",
]
addopts = "-m 'not integration'"

[tool.ruff]
# Enable flake8-bugbear (`B`) rules, and isort (`I`) for import sorting
lint.select = ["E4", "E7", "E9", "F"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Exclude a variety of commonly ignored directories
exclude = [
  ".git",
  ".eggs",
  "__pycache__",
  "build",
  "dist",
  "*.ipynb",
  "symbology/bin/",
]
# Same as Black
line-length = 160
# Target Python 3.13
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Ignore unused imports in __init__ files
"**llm*" = ["E501"]      # Ignore long lines in prompts, etc
"conftest.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["symbology"]
# Enable automatic import sorting when running Ruff
# automatically-sort-imports = true
# Ensure imports are properly organized
force-single-line = false
force-sort-within-sections = true
order-by-type = false

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["symbology*"]
